add json web token

+++ server/controllers/taskController.js
@@ -21,9 +21,10 @@ const addNewTask = async (req, res) => {
 }
 
 const getTasks = async(req, res) => {
+
   res.status(200).json({
     message: 'Tasks',
-    data: tasks
+    data: req.rule !== 0 ? []: tasks
   })
 }
 
@@ -39,7 +40,20 @@ const updateTask = async(req, res) => {
 }
 
 const deleteTask = async(req, res) => {
-  const {id} = req.query
+  try {
+    const {id} = req.query
+
+  const item = await TaskModel.findById(id)
+
+  if (!item) {
+    throw new Error('Task not found')
+  }
+
+  const createdBy = item.createdBy
+
+  if (createdBy !== req.uid) {
+    throw new Error('This not your task')
+  }
 
   await TaskModel.findByIdAndDelete(id)
 
@@ -47,11 +61,40 @@ const deleteTask = async(req, res) => {
     message: 'Removed',
     data: []
   })
+  } catch (error) {
+    res.status(404).json({
+      message: error.message,
+    })
+  }
+  
+}
+
+const getTaskDetail = async (req, res) => {
+  const {id} = req.params 
+  try {
+    const item = tasks.find(element => `${element.id}` === id)
+
+    if (!item) {
+      throw  new Error('Task not found')
+    }
+
+    res.status(200).json({
+      message: "Task detail",
+      data: item
+    })
+  } catch (error) {
+    console.log(error)
+    res.status(404).json({
+      message: error.message,
+      data: []
+    })
+  }
 }
 
 export {
   addNewTask,
   getTasks,
   updateTask,
-  deleteTask
+  deleteTask,
+  getTaskDetail
 }
